jobs routes authenticatoin completed

+++ src/controllers/job.controller.ts
@@ -3,13 +3,21 @@ import { Request, Response } from "express";
 import { inject, injectable } from "inversify";
 import { TYPES } from "../config/ioc.types";
 import { IJobService } from "../services/interfaces/ijob.service";
+import { Role } from "../../prisma/generated/prisma";
 
 @injectable()
 export default class JobController {
   constructor(@inject(TYPES.IJobService) private jobService: IJobService) {}
 
   async create(req: Request, res: Response): Promise<Response> {
-    const data = req.body;
+    const user = (req as any).user; // payload from JWT
+    if (![Role.EMPLOYER, Role.ADMIN].includes(user.role)) {
+      return res
+        .status(403)
+        .json({ message: "Forbidden: only employer or admin can create jobs" });
+    }
+
+    const data = { ...req.body, createdBy: user.id }; // auto-assign creator
     const job = await this.jobService.createJob(data);
     return res.status(201).json({ success: true, data: job });
   }
@@ -33,10 +41,19 @@ export default class JobController {
   }
 
   async delete(req: Request, res: Response): Promise<Response> {
+    const user = (req as any).user;
     const { id } = req.params;
-    const job = await this.jobService.deleteJob(id);
-    return res
-      .status(200)
-      .json({ success: true, message: "Job deleted", data: job });
+
+    const job = await this.jobService.getJobById(id);
+    if (!job) return res.status(404).json({ message: "Job not found" });
+
+    if (user.role === Role.EMPLOYER && job.createdBy !== user.id) {
+      return res
+        .status(403)
+        .json({ message: "Forbidden: cannot delete other employer's job" });
+    }
+
+    await this.jobService.deleteJob(id);
+    return res.status(200).json({ success: true, message: "Job deleted" });
   }
 }
